requirements.txt
yfinance==0.2.54
pandas==2.2.3
numpy==2.2.4
scipy==1.15.2
plotly==6.0.1
dash==3.0.4
dash-bootstrap-components==2.0.3
fastapi==0.115.12
uvicorn==0.34.0
langchain==0.3.23
langchain-aws==0.2.14
boto3==1.37.23
langgraph==0.3.21
python-dotenv==1.1.0
requests==2.32.3
